Add support for ARM64 platforms
Currently, this is only in the form of the reference implementations,
as the other ARM implementations require ACL, or SVE512 (which is not
widely implemented yet outside of a few select/recent CPUs).

ACL is not an option due to not working on Windows.

This also (sort of) cherry picks the following upstream PRs:
https://github.com/oneapi-src/oneDNN/pull/1557
https://github.com/oneapi-src/oneDNN/pull/1648
https://github.com/oneapi-src/oneDNN/pull/1654

Only the relevant sections of these have been pulled in.

This works on both Linux (gcc) and Windows (MSVC)